refactor: consolidate the queries module
